Move color input to separate ColorInputPanel component
